fix: use iconphoto and cached PhotoImage for setting window icons
Using iconbitmap on Linux only allows for 1-bit B&W XBM files to be
used, so this commits changes calls to it to calls to iconphoto
instead seeing as originally, the root window icon is already set using
a PhotoImage based off of UVR-Icon-v2.ico. We now cache the PhotoImage
that would be created as this step and use it for all windows, which gets
rid of unnecessary additional icon loads. This commit also adds a PNG version
of the icon to ensure cross-platform compatibility.